DataBaseOrm/day3-base1/orm_day3_base1/clause/clause_test.go
package clause

import (
	"reflect"
	"testing"
)

func testSelect(t *testing.T) {
	var clause Clause
	clause.Set(LIMIT, 3)
	clause.Set(SELECT, "User", []string{"*"})
	clause.Set(WHERE, "Name = ?", "Tom")
	clause.Set(ORDERBY, "Age ASC")
	sql, vars := clause.Build(SELECT, WHERE, ORDERBY, LIMIT)
	t.Log(sql, vars)
	_sql := "SELECT * FROM User WHERE Name = ?  ORDER BY Age ASC LIMIT ?"
	if sql != _sql {
		t.Fatal("failed build err sql")
	}
	if !reflect.DeepEqual(vars, []interface{}{"Tom", 3}) {
		t.Fatal("failed build err sql vars")
	}
	t.Log("ok.")
}

func TestClause_Build(t *testing.T) {
	t.Run("select", func(t *testing.T) {
		testSelect(t)
	})
}
